backend/pyproject.toml
[project]
name = "issue-multi-agent-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "pydantic-settings>=2.2.1",
  "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
